moved all genotype fb and vit tests into single file
